api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
pydantic-settings>=2.6.0
python-frontmatter>=1.1.0
pyyaml>=6.0.0
websockets>=13.0
